% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_probability.R
\name{abd_prob}
\alias{abd_prob}
\title{Reference-based error correction of amplicon sequencing data}
\usage{
abd_prob(derep, ref, error_matrix)
}
\arguments{
\item{derep}{dereplicated reads (Ns are not allowed in the reads)}

\item{ref}{the unique reference sequences of the reference seqeunces, each sequence must be in one line (Ns are not allowed in the sequences)}

\item{error_matrix}{The error matrix from the former iteration}
}
\value{
Returns the lambda value and pvalue for each reads
}
\description{
This function calculates the abundance probabilities for each reads using poisson distribution
}
\details{
Ruben Garrido-Oter's group, Plant-Microbe interaction, Max Planck Institute for Plant Breeding Research
}
\examples{

}
\author{
Pengfan Zhang
}
