% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent_error_matrix.R
\name{consis_err}
\alias{consis_err}
\title{Reference-based error correction of amplicon sequencing data}
\usage{
consis_err(fq, derep, ref, lambda_out, sampling_size, ascii, min_E=0.05, min_P=1e-40, max_diff_abs=100, max_diff_ratio=0.01)
}
\arguments{
\item{fq}{the path of the fastq file (Ns are not allowed in the reads)}

\item{derep}{the derepliacted reads by dada2 function}

\item{ref}{the reference sequences, each sequence should take up one line (Ns are not allowed in the reads)}

\item{lambda_out}{the matrix containg lambda value and pvalue from the former iteration}

\item{sampling_size}{the subsampling size of the reads}

\item{ascii}{ascii characters used to encode phred scores}

\item{min_E}{the minimum expectation value of the Possion distribution for detecting paralogs within the same strain}

\item{min_P}{the P value cutoff for identifying errouneous reads}

\item{max_diff_abs}{the maximum absolute difference in number of corrected reads between two iterations, together with max_diff_ratio, before jumping out of the iiteration}

\item{max_diff_ratio}{the maximum difference in the percentages of corrected reads between two iterations}
}
\value{
Returns the final files
}
\description{
This function iterates the error matrix till reaching the stable stage
}
\details{
Ruben Garrido-Oter's group, Plant-Microbe interaction, Max Planck Institute for Plant Breeding Research
}
\examples{

}
\author{
Pengfan Zhang
}
