% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_matrix_multicore.R
\name{error_m}
\alias{error_m}
\title{Reference-based error correction of amplicon sequencing data}
\usage{
error_m(fq, ref, sample_size, threads, ascii)
}
\arguments{
\item{fq}{the path of merged amplicon sequencing reads in fastq format (Ns are not allowed in the reads)}

\item{ref}{the unique reference sequences of the reference seqeunces, each sequence must be in one line (Ns are not allowed in the sequences)}

\item{sample_size}{the sampling size of reads to generate the transition matrix}

\item{threads}{the number of threads used to align the query reads to reference sequences}

\item{ascii}{ascii characters used to encode phred scores}
}
\value{
The output is a 20 by 43 transition probability matrix
}
\description{
This function calculate the error matrix
}
\details{
Ruben Garrido-Oter's group, Plant-Microbe interaction, Max Planck Institute for Plant Breeding Research
}
\examples{

}
\author{
Pengfan Zhang
}
