% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmer_dist}
\alias{kmer_dist}
\title{DADA2}
\usage{
kmer_dist(s1, s2, kmer_size)
}
\arguments{
\item{s1}{A \code{character(1)} of DNA sequence 1.}

\item{s2}{A \code{character(1)} of DNA sequence 2.}

\item{kmer_size}{Kmer size.}
}
\value{
The kmer distance between two sequences
}
\description{
Calculate the kmer distance between two sequences. This function is originally developed by DADA2 in 2012 and we adapt it here.
}
