% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessErr_modified.R
\name{loessErr}
\alias{loessErr}
\title{Reference-based error correction of amplicon sequencing data}
\usage{
loessErr(trans, min_err_rate=1e-07)
}
\arguments{
\item{trans}{the transition matrix}

\item{min_err_rate}{the minimum transition probability for each substitution or insertion case}
}
\value{
Returns the loess fitted error matrix
}
\description{
This function fits the loess regression to the error matrix (Parts of this script have been adapted from code originally published with DADA2 (Callahan et al., 2016).)
}
\details{
Ruben Garrido-Oter's group, Plant-Microbe interaction, Max Planck Institute for Plant Breeding Research
}
\examples{

}
\author{
Pengfan Zhang
}
