% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{trans_m}
\alias{trans_m}
\title{Reference-based error correction of amplicon sequencing data}
\usage{
trans_m(query, ascii)
}
\arguments{
\item{query}{list containing subsampled amplicon sequencing reads, quality scores, and reference sequences showing the highest identity for each read (Ns are not allowed in the reads)}

\item{ascii}{ascii characters used to encode phred scores}
}
\value{
The output is a 20 by 43 matrix containing the counts for different kinds of transitions
}
\description{
This function count the transition matrix
}
\details{
Ruben Garrido-Oter's group, Plant-Microbe interaction, Max Planck Institute for Plant Breeding Research
}
\examples{

}
\author{
Pengfan Zhang
}
